+++ src/lib/components/admin/Functions.svelte
@@ -3,7 +3,7 @@
 	import fileSaver from 'file-saver';
 	const { saveAs } = fileSaver;
 
-	import { OpenTutorAI_NAME, config, functions, models, settings } from '$lib/stores';
+	import { TUTOR_NAME, config, functions, models, settings } from '$lib/stores';
 	import { onMount, getContext, tick } from 'svelte';
 
 	import { goto } from '$app/navigation';
@@ -65,9 +65,9 @@
 			return null;
 		});
 
-		toast.success($i18n.t('Redirecting you to Open TutorAI Community'));
+		toast.success($i18n.t('Redirecting you to TutorAI Community'));
 
-		const url = 'https://openwebui.com';
+		const url = 'http://opentutorai.com/';
 
 		const tab = await window.open(`${url}/functions/create`, '_blank');
 
@@ -192,7 +192,7 @@
 
 <svelte:head>
 	<title>
-		{$i18n.t('Functions')} | {$OpenTutorAI_NAME}
+		{$i18n.t('Functions')} | {$TUTOR_NAME}
 	</title>
 </svelte:head>
 
@@ -469,12 +469,12 @@
 {#if $config?.features.enable_community_sharing}
 	<div class=" my-16">
 		<div class=" text-xl font-medium mb-1 line-clamp-1">
-			{$i18n.t('Made by Open TutorAI Community')}
+			{$i18n.t('Made by TutorAI Community')}
 		</div>
 
 		<a
 			class=" flex cursor-pointer items-center justify-between hover:bg-gray-50 dark:hover:bg-gray-850 w-full mb-2 px-3.5 py-1.5 rounded-xl transition"
-			href="https://openwebui.com/#open-webui-community"
+			href="#"
 			target="_blank"
 		>
 			<div class=" self-center">