+++ src/lib/components/playground/Completions.svelte
@@ -5,7 +5,7 @@
 	import { onMount, tick, getContext } from 'svelte';
 
 	import { TUTOR_BASE_URL } from '$lib/constants';
-	import { OpenTutorAI_NAME, config, user, models, settings, showSidebar } from '$lib/stores';
+	import { TUTOR_NAME, config, user, models, settings, showSidebar } from '$lib/stores';
 	import { chatCompletion } from '$lib/apis/openai';
 
 	import { splitStream } from '$lib/utils';