+++ src/routes/error/+page.svelte
@@ -1,6 +1,6 @@
 <script>
 	import { goto } from '$app/navigation';
-	import { OpenTutorAI_NAME, config } from '$lib/stores';
+	import { TUTOR_NAME, config } from '$lib/stores';
 	import { onMount, getContext } from 'svelte';
 
 	const i18n = getContext('i18n');
@@ -22,12 +22,12 @@
 			<div class="m-auto pb-44 flex flex-col justify-center">
 				<div class="max-w-md">
 					<div class="text-center text-2xl font-medium z-50">
-						{$i18n.t('{{OpenTutorAIName}} Backend Required', { OpenTutorAIName: $OpenTutorAI_NAME })}
+						{$i18n.t('{{OpenTutorAIName}} Backend Required', { OpenTutorAIName: $TUTOR_NAME })}
 					</div>
 
 					<div class=" mt-4 text-center text-sm w-full">
 						{$i18n.t(
-							"Oops! You're using an unsupported method (frontend only). Please serve the Open TutorAI from the backend."
+							"Oops! You're using an unsupported method (frontend only). Please serve the TutorAI from the backend."
 						)}
 
 						<br class=" " />