+++ CHANGELOG.md
@@ -15,4 +15,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- uppdating open TutorAi form open web ui.
+- uppdating TutorAI form open web ui.